Remove tenantSlug from ProductCategory create/edit endpoint models

+++ WhyNotEarth.Meredith/Shop/ProductCategoryService.cs
@@ -44,24 +44,23 @@ namespace WhyNotEarth.Meredith.Shop
 
         public async Task DeleteAsync(int id, User user)
         {
-            await CheckPermissionAsync(user, id);
-
             var category = await GetAsync(id);
+
+            if (category is null)
+            {
+                throw new RecordNotFoundException($"Category {id} not found");
+            }
+
+            await CheckPermissionAsync(user, category.Tenant.Slug);
 
             _dbContext.ProductCategories.Remove(category);
 
             await _dbContext.SaveChangesAsync();
         }
 
-        public async Task<ProductCategory> CreateAsync(ProductCategoryModel model, User user)
+        public async Task<ProductCategory> CreateAsync(string tenantSlug, ProductCategoryModel model, User user)
         {
-            var tenant = await GetTenant(model.TenantSlug!);
-            if(tenant is null)
-            {
-                throw new InvalidActionException($"Tenant {model.TenantSlug} not found");
-            }
-
-            await CheckPermissionAsync(user, tenant.Id);
+            var tenant = await CheckPermissionAsync(user, tenantSlug);
 
             var category = new ProductCategory
             {
@@ -76,15 +75,9 @@ namespace WhyNotEarth.Meredith.Shop
             return category;
         }
 
-        public async Task<ProductCategory> EditAsync(int categoryId, ProductCategoryModel model, User user)
+        public async Task<ProductCategory> EditAsync(string tenantSlug, int categoryId, ProductCategoryModel model, User user)
         {
-            var tenant = await GetTenant(model.TenantSlug!);
-            if (tenant is null)
-            {
-                throw new InvalidActionException($"Tenant {model.TenantSlug} not found");
-            }
-
-            await CheckPermissionAsync(user, tenant.Id);
+            await CheckPermissionAsync(user, tenantSlug);
 
             var category = await _dbContext.ProductCategories
                 .FirstOrDefaultAsync(item => item.Id == categoryId);
@@ -103,20 +96,22 @@ namespace WhyNotEarth.Meredith.Shop
             return category;
         }
 
-        private async Task<Data.Entity.Models.Tenant> GetTenant(string tenantSlug)
+        private async Task<Data.Entity.Models.Tenant> CheckPermissionAsync(User user, string tenantSlug)
         {
-            return await _dbContext.Tenants.FirstOrDefaultAsync(item =>
-                item.Slug == tenantSlug);
-        }
+            var tenant = await _dbContext.Tenants.FirstOrDefaultAsync(item => item.Slug == tenantSlug && item.OwnerId == user.Id);
 
-        private async Task CheckPermissionAsync(User user, int tenantId)
-        {
-            var ownsTenant = await _dbContext.Tenants.AnyAsync(item => item.Id == tenantId && item.OwnerId == user.Id);
-
-            if (!ownsTenant)
+            if (tenant is null)
             {
+                tenant = await _dbContext.Tenants.FirstOrDefaultAsync(item => item.Slug == tenantSlug);
+                if (tenant is null)
+                {
+                    throw new RecordNotFoundException($"Tenant {tenantSlug} not found");
+                }
+
                 throw new ForbiddenException("You don't own this tenant");
             }
+
+            return tenant;
         }
     }
 }
